+++ tools/external_tools.go
@@ -8,4 +8,5 @@ package main
 import (
 	_ "github.com/client9/misspell/cmd/misspell"
 	_ "github.com/golangci/golangci-lint/cmd/golangci-lint"
+	_ "golang.org/x/exp/cmd/gorelease"
 )